reorg(routing): move api/_meta.py -> core/routing.py
Auth and RouterMeta are shared mount primitives, not app-shell
internals: the 42 domain routers import them, and after #488 those
routers live under core/<domain>/. Put the primitives in a core/ leaf so
a colocated router imports up the layer (core -> core), never back into
the api package.

Pure move + repoint: 42 routers, _discovery, and the discovery test now
import `from core.routing import Auth, RouterMeta`. Discovery still scans
backend/api/ and mounts 366 route-method pairs unchanged; 13/13
tests/unit/test_router_discovery.py pass.
