+++ backend/api/analytics.py
@@ -22,7 +22,7 @@ from core.storage.models import Finding, Case, CaseClosureInfo, LLMInteractionLo
 from core.storage.connection import get_db, get_db_session
 from backend.services.ai_insights_service import AIInsightsService
 from core.threat_intel.mitre_lookup import get_time_range, resolve_technique  # noqa: F401
-from api._meta import Auth, RouterMeta
+from core.routing import Auth, RouterMeta
 
 logger = logging.getLogger(__name__)
