+++ backend/api/config.py
@@ -10,7 +10,7 @@ import os
 
 # Import new secrets manager
 import sys
-from api._meta import Auth, RouterMeta
+from core.routing import Auth, RouterMeta
 
 sys.path.insert(0, str(Path(__file__).parent.parent))
 from secrets_manager import get_secret, set_secret, delete_secret, get_secrets_manager
