+++ backend/api/llm_providers.py
@@ -18,7 +18,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from pydantic import BaseModel, Field
 from sqlalchemy import delete as sa_delete, update
 from sqlalchemy.orm import Session
-from api._meta import Auth, RouterMeta
+from core.routing import Auth, RouterMeta
 
 sys.path.insert(0, str(Path(__file__).parent.parent))
 from secrets_manager import delete_secret, get_secret, set_secret  # noqa: E402
