+++ backend/api/storage_status.py
@@ -11,7 +11,7 @@ from fastapi import APIRouter, Depends, HTTPException
 
 from backend.middleware.auth import get_current_active_user, require_settings_admin
 from core.storage.models import User
-from api._meta import Auth, RouterMeta
+from core.routing import Auth, RouterMeta
 
 logger = logging.getLogger(__name__)
